Delete legacy v2 question execution code
